requirements/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
discord.py==2.3.2
Flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
priority==2.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
Quart==0.19.4
regex==2023.10.3
requests==2.31.0
six==1.16.0
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
